debug grammar explain
