+++ app/config.py
@@ -1,6 +1,6 @@
-TOKEN=""
+TOKEN="7797670938:AAFPhQORQia5S7ePrFkNd2OdGNgi0L0Dxl0"
 TEST_LEVEL_API_KEY=""
-EXPLATION_GRAMMAR_RULES_API_KEY=""
-GRAMMAR_EXERCISES_API_KEY=""
-DAILY_WORDS_API_KEY=""
-TRANSLATE_API_KEY=""
+EXPLATION_GRAMMAR_RULES_API_KEY="tjNW0ju7WlPdGqnlqKba355xMdsP4jGJ"
+GRAMMAR_EXERCISES_API_KEY="Cnbpbf0qkF6jyb9nXW9moIhLkKAwCOWx"
+DAILY_WORDS_API_KEY="mBZxsljeEsG3dp9oaDNfil6PgpEj2TIf"
+TRANSLATE_API_KEY="ZTx46YXBmHjkHkUPTIAUVr0FEfz0QELj"
