+++ app/database_api/database.json
@@ -1,10 +1,10 @@
 {
-    "2030231313": {
-        "language": "Немецкий",
-        "level": "Продвинутый"
-    },
     "1454790373": {
         "language": "Итальянский",
         "level": "Начианющий"
+    },
+    "2030231313": {
+        "language": "Английский",
+        "level": "Начинающий"
     }
 }
