+++ app/keyboards/modules_select_kb.py
@@ -2,12 +2,12 @@ from aiogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 
 
 modules_inline = InlineKeyboardMarkup(inline_keyboard=[
-    [InlineKeyboardButton(text="Переводчик", callback_data="translate no")],
-    [InlineKeyboardButton(text="Упражнения", callback_data="exercise no")],
-    [InlineKeyboardButton(text="Объяснения", callback_data="explation no")],
-    [InlineKeyboardButton(text="Назад⬛️", callback_data="back_to_level_select no")]
+    [InlineKeyboardButton(text="Переводчик", callback_data="translate")],
+    [InlineKeyboardButton(text="Упражнения", callback_data="exercise")],
+    [InlineKeyboardButton(text="Объяснения", callback_data="explation")],
+    [InlineKeyboardButton(text="Назад⬛️", callback_data="back_to_level_select")]
 ])
 
 back_to_modules = InlineKeyboardMarkup(inline_keyboard=[
-    [InlineKeyboardButton(text="Назад⬛️", callback_data="back_to_modules no")]
+    [InlineKeyboardButton(text="Назад⬛️", callback_data="back_to_modules")]
 ])
